2-Code/requirements.txt
requests
python-dotenv
flask
flask-pymongo
flask-cors
pytesseract
pillow
azure-storage-blob

# Ajoutez d'autres dépendances ici si nécessaire
